Split out db.rs into more files; Bot now auto leaves voice channel if only one left
